functions/pandas/requirements.txt

google-cloud-firestore
pandas
numpy
faker
anonymizedf
sendgrid
